reinitialized db migrations
