+++ alembic/versions/347b7fa43720_first_revision.py
@@ -1,0 +1,24 @@
+"""first revision
+
+Revision ID: 347b7fa43720
+Revises: 
+Create Date: 2021-11-15 12:35:36.844593
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '347b7fa43720'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass
